Refactor: set dashboard as a page
